Edits in weather client

+++ lib/features/data/datasource/weather_client/models.dart
@@ -1,3 +1,5 @@
+// The file implements DTO models
+
 import 'dart:core';
 
 import 'package:flutter_weather/common/utils.dart';
@@ -6,7 +8,7 @@ class CurrentWeatherDTO {
   late String name;
   late double temp;
   late double wind;
-  late int humidity;
+  late double humidity;
 
   CurrentWeatherDTO(
       {required this.name,
@@ -20,7 +22,7 @@ class CurrentWeatherDTO {
     name = json['location']['name'];
     temp = current['temp_c'];
     wind = kilometerPerHourToMeterPerSecond(current['wind_kph']);
-    humidity = current['humidity'];
+    humidity = (current['humidity'] as int).toDouble();
   }
 }
